renamed to sparkstorecommand.py and turned into linemagic
