+++ sparkmagic/sparkmagic/livyclientlib/sparkstorecommand.py
