+++ sparkmagic/sparkmagic/tests/test_remotesparkmagics.py
@@ -8,7 +8,7 @@ from sparkmagic.livyclientlib.command import Command
 from sparkmagic.livyclientlib.endpoint import Endpoint
 from sparkmagic.livyclientlib.exceptions import *
 from sparkmagic.livyclientlib.sqlquery import SQLQuery
-from sparkmagic.livyclientlib.sparkcommand import SparkStoreCommand
+from sparkmagic.livyclientlib.sparkstorecommand import SparkStoreCommand
 
 magic = None
 spark_controller = None
